move cometpods out of copepod folder
