+++ Assets/Textures/NPCs/CosmostoneShowers/TerritorialCometpod.png
